requirements.txt
numpy
scipy
tensorflow-gpu==2.5.0
tensorboard==2.5.0
pillow==8.2.0
opencv-python==4.5.3.56